TICKET-145: split lobby.tsx into focused modules
Extract WebRTC handshake logic into network/webrtc.ts, username
persistence into username.ts, and shared UI helpers into
ui/lobbyHelpers.tsx. Lobby screens remain in lobby.tsx but import
from extracted modules. Fix jest moduleNameMapper to resolve
@pulse-ts/core and @pulse-ts/network for isolated test runs.
