+++ demos/arena/src/lobby.tsx
@@ -1,43 +1,22 @@
-import { createElement, Column, Row, Button } from '@pulse-ts/dom';
 import {
     applyStaggeredEntrance,
     applyButtonHoverScale,
 } from './overlayAnimations';
-import { DataChannelTransport } from '@pulse-ts/network/transports/datachannel';
 import type { Transport } from '@pulse-ts/network';
 import { getAppVersion, isUpdateAvailable } from './versionCheck';
 import { versionsMatch } from './versionMatch';
-
-/**
- * Fallback ICE servers (STUN only) used when TURN credentials
- * are unavailable or the signaling server doesn't support them.
- */
-const FALLBACK_ICE_SERVERS: RTCIceServer[] = [
-    { urls: 'stun:stun.l.google.com:19302' },
-    { urls: 'stun:stun1.l.google.com:19302' },
-];
-
-/** Timeout (ms) waiting for WebRTC DataChannel to open. */
-const WEBRTC_TIMEOUT = 15000;
-
-/** Timeout (ms) waiting for ICE server credentials from signaling server. */
-const ICE_SERVERS_TIMEOUT = 5000;
-
-/**
- * Build the Lambda signaling WebSocket URL.
- * In production, this comes from the Terraform output (API Gateway WebSocket endpoint).
- * For local development, falls back to localhost.
- */
-function getSignalingUrl(): string {
-    const envUrl =
-        typeof (window as any).__SIGNALING_URL__ === 'string'
-            ? (window as any).__SIGNALING_URL__
-            : undefined;
-    if (envUrl) return envUrl;
-
-    const proto = window.location.protocol === 'https:' ? 'wss:' : 'ws:';
-    return `${proto}//${window.location.host}`;
-}
+import { getSignalingUrl, requestIceServers, establishP2P } from './network/webrtc';
+import { getUsername, setUsername, hasUsername, USERNAME_MAX_LENGTH } from './username';
+import {
+    createOverlay,
+    clearAndCreateContent,
+    createHeading,
+    createSubheading,
+    createBtn,
+    createColumnEl,
+    createRowEl,
+    createStatusIndicator,
+} from './ui/lobbyHelpers';
 
 /** Result returned when the host completes the lobby. */
 export interface HostResult {
@@ -95,222 +74,6 @@ export function showLobby(
         }
 
         showLobbyMenu(overlay, finish);
-    });
-}
-
-// ---------------------------------------------------------------------------
-// WebRTC Handshake
-// ---------------------------------------------------------------------------
-
-/**
- * Request TURN relay credentials from the signaling server via the
- * `get-ice-servers` action. Falls back to STUN-only if unavailable.
- *
- * @param ws - Open WebSocket to the signaling server.
- * @returns ICE server configuration including TURN credentials.
- */
-function requestIceServers(ws: WebSocket): Promise<RTCIceServer[]> {
-    return new Promise((resolve) => {
-        const timeout = setTimeout(() => {
-            ws.removeEventListener('message', onMessage);
-            resolve(FALLBACK_ICE_SERVERS);
-        }, ICE_SERVERS_TIMEOUT);
-
-        function onMessage(event: MessageEvent) {
-            let msg: any;
-            try {
-                msg = JSON.parse(
-                    typeof event.data === 'string'
-                        ? event.data
-                        : new TextDecoder().decode(event.data),
-                );
-            } catch {
-                return;
-            }
-            if (msg.type !== 'ice-servers') return;
-
-            clearTimeout(timeout);
-            ws.removeEventListener('message', onMessage);
-
-            const servers: RTCIceServer[] = [
-                ...FALLBACK_ICE_SERVERS,
-                ...(msg.iceServers ?? []),
-            ];
-            resolve(servers);
-        }
-
-        ws.addEventListener('message', onMessage);
-        ws.send(JSON.stringify({ action: 'get-ice-servers' }));
-    });
-}
-
-/**
- * Establish a WebRTC P2P connection via the Lambda signaling server.
- *
- * @param signalingWs - Open WebSocket to the signaling server.
- * @param isHost - Whether this peer is the offerer (host creates DataChannel).
- * @param peerConnectionId - The signaling connectionId of the remote peer.
- * @param iceServers - ICE server configuration (STUN + TURN).
- * @returns Promise resolving with the DataChannelTransport once the DataChannel opens.
- */
-function establishP2P(
-    signalingWs: WebSocket,
-    isHost: boolean,
-    peerConnectionId: string,
-    iceServers: RTCIceServer[],
-    bufferedMessages: MessageEvent[] = [],
-): Promise<DataChannelTransport> {
-    const role = isHost ? 'HOST' : 'JOINER';
-    return new Promise((resolve, reject) => {
-        let pc: RTCPeerConnection;
-        try {
-            pc = new RTCPeerConnection({ iceServers });
-        } catch (e: any) {
-            reject(e);
-            return;
-        }
-        let dc: RTCDataChannel | null = null;
-        let resolved = false;
-
-        const timeout = setTimeout(() => {
-            if (!resolved) {
-                resolved = true;
-                try {
-                    pc.close();
-                } catch {}
-                reject(new Error('WebRTC connection timed out'));
-            }
-        }, WEBRTC_TIMEOUT);
-
-        function done(channel: RTCDataChannel) {
-            if (resolved) return;
-            resolved = true;
-            clearTimeout(timeout);
-            resolve(new DataChannelTransport(channel, pc, peerConnectionId));
-        }
-
-        pc.onicecandidate = (ev) => {
-            if (ev.candidate) {
-                signalingWs.send(
-                    JSON.stringify({
-                        action: 'signal',
-                        data: { type: 'ice', candidate: ev.candidate },
-                    }),
-                );
-            }
-        };
-
-        const pendingIceCandidates: RTCIceCandidate[] = [];
-        let remoteDescriptionSet = false;
-
-        function flushIceCandidates() {
-            remoteDescriptionSet = true;
-            for (const c of pendingIceCandidates) {
-                pc.addIceCandidate(c).catch((err) =>
-                    console.error(
-                        `[P2P ${role}] Error adding queued ICE candidate:`,
-                        err,
-                    ),
-                );
-            }
-            pendingIceCandidates.length = 0;
-        }
-
-        const onSignal = (event: MessageEvent) => {
-            let msg: any;
-            try {
-                msg = JSON.parse(
-                    typeof event.data === 'string'
-                        ? event.data
-                        : new TextDecoder().decode(event.data),
-                );
-            } catch {
-                return;
-            }
-
-            if (msg.type !== 'signal') return;
-            const signal = msg.data;
-
-            if (signal.type === 'offer' && signal.sdp) {
-                pc.setRemoteDescription(new RTCSessionDescription(signal))
-                    .then(() => {
-                        flushIceCandidates();
-                        return pc.createAnswer();
-                    })
-                    .then((answer) => pc.setLocalDescription(answer))
-                    .then(() => {
-                        signalingWs.send(
-                            JSON.stringify({
-                                action: 'signal',
-                                data: pc.localDescription,
-                            }),
-                        );
-                    })
-                    .catch((err) =>
-                        console.error(
-                            `[P2P ${role}] Error handling offer:`,
-                            err,
-                        ),
-                    );
-            } else if (signal.type === 'answer' && signal.sdp) {
-                pc.setRemoteDescription(new RTCSessionDescription(signal))
-                    .then(() => flushIceCandidates())
-                    .catch((err) =>
-                        console.error(
-                            `[P2P ${role}] Error handling answer:`,
-                            err,
-                        ),
-                    );
-            } else if (signal.type === 'ice' && signal.candidate) {
-                const candidate = new RTCIceCandidate(signal.candidate);
-                if (remoteDescriptionSet) {
-                    pc.addIceCandidate(candidate).catch((err) =>
-                        console.error(
-                            `[P2P ${role}] Error adding ICE candidate:`,
-                            err,
-                        ),
-                    );
-                } else {
-                    pendingIceCandidates.push(candidate);
-                }
-            }
-        };
-
-        signalingWs.addEventListener('message', onSignal);
-
-        for (const msg of bufferedMessages) {
-            onSignal(msg);
-        }
-
-        if (isHost) {
-            dc = pc.createDataChannel('game', { ordered: false });
-            dc.binaryType = 'arraybuffer';
-            dc.onopen = () => done(dc!);
-
-            pc.createOffer()
-                .then((offer) => pc.setLocalDescription(offer))
-                .then(() => {
-                    signalingWs.send(
-                        JSON.stringify({
-                            action: 'signal',
-                            data: pc.localDescription,
-                        }),
-                    );
-                })
-                .catch((err) => {
-                    console.error(`[P2P ${role}] Error creating offer:`, err);
-                });
-        } else {
-            pc.ondatachannel = (ev) => {
-                dc = ev.channel;
-                dc.binaryType = 'arraybuffer';
-                if (dc.readyState === 'open') {
-                    done(dc);
-                } else {
-                    dc.onopen = () => done(dc!);
-                }
-            };
-        }
     });
 }
 
@@ -879,179 +642,4 @@ function showVersionMismatch(
             showLobbyMenu(overlay, finish);
         });
     }
-}
-
-// ---------------------------------------------------------------------------
-// Username management
-// ---------------------------------------------------------------------------
-
-const USERNAME_KEY = 'pulse-arena-username';
-const USERNAME_MAX_LENGTH = 24;
-
-function getUsername(): string {
-    try {
-        return localStorage.getItem(USERNAME_KEY) || '';
-    } catch {
-        return '';
-    }
-}
-
-function hasUsername(): boolean {
-    return getUsername().length > 0;
-}
-
-function setUsername(name: string): void {
-    const validated = name.trim().slice(0, USERNAME_MAX_LENGTH);
-    try {
-        localStorage.setItem(USERNAME_KEY, validated);
-    } catch {
-        /* localStorage unavailable */
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Shared UI helpers
-// ---------------------------------------------------------------------------
-
-function createOverlay(): HTMLDivElement {
-    const { root } = createElement(
-        <div
-            style={{
-                position: 'absolute',
-                inset: '0',
-                zIndex: '5000',
-                display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'center',
-                backgroundColor: 'rgba(10, 10, 26, 0.65)',
-                opacity: '0',
-                transition: 'opacity 0.4s ease-in-out',
-            }}
-        />,
-    );
-    return root as HTMLDivElement;
-}
-
-function clearAndCreateContent(overlay: HTMLElement): HTMLDivElement {
-    overlay.innerHTML = '';
-    const { root } = createElement(
-        <Column
-            gap={16}
-            style={{
-                alignItems: 'center',
-                padding: '0 20px',
-            }}
-        />,
-    );
-    const content = root as HTMLDivElement;
-    overlay.appendChild(content);
-    return content;
-}
-
-function createHeading(text: string): HTMLDivElement {
-    const { root } = createElement(
-        <div
-            style={{
-                font: 'bold clamp(22px, 6vw, 32px) monospace',
-                color: '#fff',
-                textShadow: '0 0 16px rgba(72, 201, 176, 0.5)',
-                letterSpacing: '3px',
-                marginBottom: '8px',
-            }}
-        >
-            {text}
-        </div>,
-    );
-    return root as HTMLDivElement;
-}
-
-function createSubheading(text: string): HTMLDivElement {
-    const { root } = createElement(
-        <div
-            style={{
-                font: '14px monospace',
-                color: '#888',
-                letterSpacing: '2px',
-                textTransform: 'uppercase',
-            }}
-        >
-            {text}
-        </div>,
-    );
-    return root as HTMLDivElement;
-}
-
-function createBtn(label: string, color: string): HTMLButtonElement {
-    const { root } = createElement(
-        <Button
-            accent={color}
-            style={{
-                font: 'bold clamp(14px, 3.5vw, 18px) monospace',
-                border: '2px solid rgba(255,255,255,0.2)',
-                borderRadius: '6px',
-                padding: '12px 32px',
-                minWidth: 'min(200px, 70vw)',
-                minHeight: '44px',
-            }}
-        >
-            {label}
-        </Button>,
-    );
-    return root as HTMLButtonElement;
-}
-
-function createColumnEl(...children: HTMLElement[]): HTMLDivElement {
-    const { root } = createElement(
-        <Column gap={12} style={{ alignItems: 'center' }} />,
-    );
-    const col = root as HTMLDivElement;
-    children.forEach((c) => col.appendChild(c));
-    return col;
-}
-
-function createRowEl(...children: HTMLElement[]): HTMLDivElement {
-    const { root } = createElement(
-        <Row gap={12} center style={{ flexWrap: 'wrap' }} />,
-    );
-    const row = root as HTMLDivElement;
-    children.forEach((c) => row.appendChild(c));
-    return row;
-}
-
-/** Status indicator element with a setter for status text and color. */
-interface StatusIndicator {
-    el: HTMLDivElement;
-    set: (
-        state: 'idle' | 'connecting' | 'connected' | 'error',
-        msg?: string,
-    ) => void;
-}
-
-const STATUS_COLORS: Record<string, string> = {
-    idle: '#888',
-    connecting: '#f1c40f',
-    connected: '#48c9b0',
-    error: '#e74c3c',
-};
-
-function createStatusIndicator(): StatusIndicator {
-    const { root } = createElement(
-        <div
-            style={{
-                font: '13px monospace',
-                color: '#888',
-                minHeight: '20px',
-                textAlign: 'center',
-            }}
-        />,
-    );
-    const el = root as HTMLDivElement;
-
-    return {
-        el,
-        set(state, msg) {
-            el.style.color = STATUS_COLORS[state] ?? '#888';
-            el.textContent = msg ?? '';
-        },
-    };
 }
